Update the files

+++ utils.py
@@ -4,10 +4,10 @@ math_function_regex=re.compile("(?:[0-9-+*/()x]|pi|e)+") #a simple regex that re
 
 colorstring = 'bgkry'
 
-def binary_search(arr, x: float):
-    low=0
-    high=len(arr)-1
-    mid=0
+def binary_search(arr, x: int):
+    low = 0
+    high = len(arr) - 1
+    mid = 0
     while low <= high:
         mid = (high + low) // 2
         if arr[mid].end < x:
@@ -15,14 +15,14 @@ def binary_search(arr, x: float):
         elif arr[mid].start > x:
             high = mid - 1
         else:
-            if arr[mid].start<x and x<arr[mid].end:
+            if arr[mid].start < x and x < arr[mid].end:
                 return mid
             elif arr[mid].start == x:
                 if arr[mid].equal[0]: #if x is a boarder we have to ensure that if within functions domain
                     return mid
                 else:
                     high = mid - 1
-            elif arr[mid].end==x:
+            elif arr[mid].end == x:
                 if arr[mid].equal[1]:
                     return mid
                 else:
